+++ tests/schema/test_generate_schema.py
@@ -1,7 +1,6 @@
 # pylint: disable=C0301
 import re
-
-from asynctest.mock import CoroutineMock, patch
+from unittest.mock import AsyncMock, patch
 
 from tortoise import Tortoise
 from tortoise.contrib import test
@@ -30,7 +29,7 @@ class TestGenerateSchema(test.SimpleTestCase):
 
     async def init_for(self, module: str, safe=False) -> None:
         with patch(
-            "tortoise.backends.sqlite.client.SqliteClient.create_connection", new=CoroutineMock()
+            "tortoise.backends.sqlite.client.SqliteClient.create_connection", new=AsyncMock()
         ):
             await Tortoise.init(
                 {
@@ -396,7 +395,7 @@ CREATE TABLE "team_team" (
 class TestGenerateSchemaMySQL(TestGenerateSchema):
     async def init_for(self, module: str, safe=False) -> None:
         try:
-            with patch("asyncmy.create_pool", new=CoroutineMock()):
+            with patch("asyncmy.create_pool", new=AsyncMock()):
                 await Tortoise.init(
                     {
                         "connections": {
@@ -776,7 +775,7 @@ CREATE TABLE `team_team` (
 class TestGenerateSchemaPostgresSQL(TestGenerateSchema):
     async def init_for(self, module: str, safe=False) -> None:
         try:
-            with patch("asyncpg.create_pool", new=CoroutineMock()):
+            with patch("asyncpg.create_pool", new=AsyncMock()):
                 await Tortoise.init(
                     {
                         "connections": {