+++ CHANGELOG.rst
@@ -18,7 +18,11 @@ Changed
 ^^^^^^^
 - Move `Function`, `Aggregate` from `functions.py` to `expressions.py`. (#943)
 - Move `Q` from `query_utils.py` to `expressions.py`.
-
+Removed
+^^^^^^^
+- Remove `asynctest` and use `unittest.IsolatedAsyncioTestCase`. (#416)
+- Remove `py37` support in tests.
+- Remove `green` and `nose2` test runner.
 0.17
 ====
 0.17.8